Updated `airports.source.schema.graphql` schema to reflect changes made in schema generation logic and updated tests accordingly. Changed Case05 and Case07 to execute all app sync queries instead of just getNodeContinents. Added check to skip the queryAppSync.test.js if no env vars are set.
