+++ test/TestCases/Case07/outputReference/output.source.schema.graphql
@@ -1,17 +1,43 @@
+enum SortingDirection {
+  ASC
+  DESC
+}
+
 type Continent @alias(property: "continent") {
   id: ID! @id
   code: String
   type: String
   desc: String
-  airportContainssOut(filter: AirportInput, options: Options): [Airport] @relationship(edgeType: "contains", direction: OUT)
+  airportContainssOut(filter: AirportInput, options: Options, sort: [AirportSort!]): [Airport] @relationship(edgeType: "contains", direction: OUT)
   contains: Contains
 }
 
 input ContinentInput {
   id: ID @id
+  code: StringScalarFilters
+  type: StringScalarFilters
+  desc: StringScalarFilters
+}
+
+input ContinentCreateInput {
+  id: ID @id
   code: String
   type: String
   desc: String
+}
+
+input ContinentUpdateInput {
+  id: ID! @id
+  code: String
+  type: String
+  desc: String
+}
+
+input ContinentSort {
+  id: SortingDirection
+  type: SortingDirection
+  code: SortingDirection
+  desc: SortingDirection
 }
 
 type Country @alias(property: "country") {
@@ -19,15 +45,36 @@ type Country @alias(property: "country") {
   code: String
   type: String
   desc: String
-  airportContainssOut(filter: AirportInput, options: Options): [Airport] @relationship(edgeType: "contains", direction: OUT)
+  airportContainssOut(filter: AirportInput, options: Options, sort: [AirportSort!]): [Airport] @relationship(edgeType: "contains", direction: OUT)
   contains: Contains
 }
 
 input CountryInput {
   _id: ID @id
+  code: StringScalarFilters
+  type: StringScalarFilters
+  desc: StringScalarFilters
+}
+
+input CountryCreateInput {
+  _id: ID @id
   code: String
   type: String
   desc: String
+}
+
+input CountryUpdateInput {
+  _id: ID! @id
+  code: String
+  type: String
+  desc: String
+}
+
+input CountrySort {
+  _id: SortingDirection
+  code: SortingDirection
+  type: SortingDirection
+  desc: SortingDirection
 }
 
 type Version @alias(property: "version") {
@@ -41,11 +88,38 @@ type Version @alias(property: "version") {
 
 input VersionInput {
   _id: ID @id
+  date: StringScalarFilters
+  code: StringScalarFilters
+  author: StringScalarFilters
+  type: StringScalarFilters
+  desc: StringScalarFilters
+}
+
+input VersionCreateInput {
+  _id: ID @id
   date: String
   code: String
   author: String
   type: String
   desc: String
+}
+
+input VersionUpdateInput {
+  _id: ID! @id
+  date: String
+  code: String
+  author: String
+  type: String
+  desc: String
+}
+
+input VersionSort {
+  _id: SortingDirection
+  date: SortingDirection
+  code: SortingDirection
+  author: SortingDirection
+  type: SortingDirection
+  desc: SortingDirection
 }
 
 type Airport @alias(property: "airport") {
@@ -65,13 +139,29 @@ type Airport @alias(property: "airport") {
   outboundRoutesCount: Int @graphQuery(statement: "MATCH (this)-[r:route]->(a) RETURN count(r)")
   continentContainsIn: Continent @relationship(edgeType: "contains", direction: IN)
   countryContainsIn: Country @relationship(edgeType: "contains", direction: IN)
-  airportRoutesOut(filter: AirportInput, options: Options): [Airport] @relationship(edgeType: "route", direction: OUT)
-  airportRoutesIn(filter: AirportInput, options: Options): [Airport] @relationship(edgeType: "route", direction: IN)
+  airportRoutesOut(filter: AirportInput, options: Options, sort: [AirportSort!]): [Airport] @relationship(edgeType: "route", direction: OUT)
+  airportRoutesIn(filter: AirportInput, options: Options, sort: [AirportSort!]): [Airport] @relationship(edgeType: "route", direction: IN)
   contains: Contains
   route: Route
 }
 
 input AirportInput {
+  _id: ID @id
+  country: StringScalarFilters
+  longest: Int
+  code: StringScalarFilters
+  city: StringScalarFilters
+  elev: Int
+  icao: StringScalarFilters
+  lon: Float
+  runways: Int
+  region: StringScalarFilters
+  type: StringScalarFilters
+  lat: Float
+  desc: StringScalarFilters
+}
+
+input AirportCreateInput {
   _id: ID @id
   country: String
   longest: Int
@@ -85,6 +175,37 @@ input AirportInput {
   type: String
   lat: Float
   desc: String
+}
+
+input AirportUpdateInput {
+  _id: ID! @id
+  country: String
+  longest: Int
+  code: String
+  city: String
+  elev: Int
+  icao: String
+  lon: Float
+  runways: Int
+  region: String
+  type: String
+  lat: Float
+  desc: String
+}
+
+input AirportSort {
+  country: SortingDirection
+  longest: SortingDirection
+  code: SortingDirection
+  city: SortingDirection
+  elev: SortingDirection
+  icao: SortingDirection
+  lon: SortingDirection
+  runways: SortingDirection
+  region: SortingDirection
+  type: SortingDirection
+  lat: SortingDirection
+  desc: SortingDirection
 }
 
 type Contains @alias(property: "contains") {
@@ -102,6 +223,14 @@ input RouteInput {
 
 input Options {
   limit: Int
+  offset: Int
+}
+
+input StringScalarFilters {
+  eq: String
+  contains: String
+  endsWith: String
+  startsWith: String
 }
 
 type Query {
@@ -111,30 +240,30 @@ type Query {
   getContinentsWithGremlin: [Continent] @graphQuery(statement: "g.V().hasLabel('continent').elementMap().fold()")
   getCountriesCountGremlin: Int @graphQuery(statement: "g.V().hasLabel('country').count()")
   getNodeContinent(filter: ContinentInput): Continent
-  getNodeContinents(filter: ContinentInput, options: Options): [Continent]
+  getNodeContinents(filter: ContinentInput, options: Options, sort: [ContinentSort!]): [Continent]
   getNodeCountry(filter: CountryInput): Country
-  getNodeCountrys(filter: CountryInput, options: Options): [Country]
+  getNodeCountrys(filter: CountryInput, options: Options, sort: [CountrySort!]): [Country]
   getNodeVersion(filter: VersionInput): Version
-  getNodeVersions(filter: VersionInput, options: Options): [Version]
+  getNodeVersions(filter: VersionInput, options: Options, sort: [VersionSort!]): [Version]
   getNodeAirport(filter: AirportInput): Airport
-  getNodeAirports(filter: AirportInput, options: Options): [Airport]
+  getNodeAirports(filter: AirportInput, options: Options, sort: [AirportSort!]): [Airport]
 }
 
 type Mutation {
   createAirport(input: AirportInput!): Airport @graphQuery(statement: "CREATE (this:airport {$input}) RETURN this")
   addRoute(fromAirportCode: String, toAirportCode: String, dist: Int): Route @graphQuery(statement: "MATCH (from:airport{code:'$fromAirportCode'}), (to:airport{code:'$toAirportCode'}) CREATE (from)-[this:route{dist:$dist}]->(to) RETURN this")
   deleteAirport(id: ID): Int @graphQuery(statement: "MATCH (this:airport) WHERE ID(this) = '$id' DETACH DELETE this")
-  createNodeContinent(input: ContinentInput!): Continent
-  updateNodeContinent(input: ContinentInput!): Continent
+  createNodeContinent(input: ContinentCreateInput!): Continent
+  updateNodeContinent(input: ContinentUpdateInput!): Continent
   deleteNodeContinent(_id: ID!): Boolean
-  createNodeCountry(input: CountryInput!): Country
-  updateNodeCountry(input: CountryInput!): Country
+  createNodeCountry(input: CountryCreateInput!): Country
+  updateNodeCountry(input: CountryUpdateInput!): Country
   deleteNodeCountry(_id: ID!): Boolean
-  createNodeVersion(input: VersionInput!): Version
-  updateNodeVersion(input: VersionInput!): Version
+  createNodeVersion(input: VersionCreateInput!): Version
+  updateNodeVersion(input: VersionUpdateInput!): Version
   deleteNodeVersion(_id: ID!): Boolean
-  createNodeAirport(input: AirportInput!): Airport
-  updateNodeAirport(input: AirportInput!): Airport
+  createNodeAirport(input: AirportCreateInput!): Airport
+  updateNodeAirport(input: AirportUpdateInput!): Airport
   deleteNodeAirport(_id: ID!): Boolean
   connectNodeContinentToNodeAirportEdgeContains(from_id: ID!, to_id: ID!): Contains
   deleteEdgeContainsFromContinentToAirport(from_id: ID!, to_id: ID!): Boolean
